- Move to new Azure account - Support CORS

+++ CL.Web/CL.Web/Scripts/cl/asset/view-item.js
@@ -28,13 +28,18 @@
     },
 
     destroy: function () {
-        var self = this;
-        this.model.destroy({
-            success: function () {
-                self.render();
-                console.log("Delete: " + self.model.get('id'));
-            }
-        });
+        if (confirm("Are you sure you'd like to delete this item?"))
+        {
+            var self = this;
+
+            this.model.destroy({
+                success: function () {
+                    self.render();
+                    console.log("Delete: " + self.model.get('id'));
+                }
+            });
+        }
+
     },
 
     render: function () {
